admin add event , delete event , update event , delete photo functionality added with token security , rendering blob files to frontend working

+++ backend_api/music_app/routers/admin.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Query, Body, Path, Header, Depends, HTTPException, status , File , UploadFile , Form
+from fastapi import APIRouter, Query, Body, Path, Header, Depends, HTTPException, status , File , UploadFile , Form , Response
 from fastapi.encoders import jsonable_encoder
 from fastapi.responses import HTMLResponse
 from typing import List, Dict
@@ -28,7 +28,7 @@ oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/admin/token")
 
 SECRET_KEY = "09d25e094faa6ca2556c818166b7a9563b93f7099f6f0f4caa6cf63b88e8d3e7"
 ALGORITHM = "HS256"
-ACCESS_TOKEN_EXPIRE_MINUTES = 15
+ACCESS_TOKEN_EXPIRE_MINUTES = 100
 def get_db():
     db = SessionLocal()
     try:
@@ -123,8 +123,14 @@ async def make_new_user(* , db : Session = Depends(get_db) , new_user : schemas.
 
 
 @router.post('/addEvent')
-async def add_event(newEvent : schemas.EventCreate = Body(...) , db : Session = Depends(get_db) , token : str = Body(...)):
-    payload = jwt.decode(token , SECRET_KEY , algorithms=[ALGORITHM])
+async def add_event(*,newEvent : schemas.EventCreate = Body(...) , db : Session = Depends(get_db) , token : str = Body(...) , res : Response):
+    
+    try:
+        payload = jwt.decode(token , SECRET_KEY , algorithms=[ALGORITHM])
+    except:
+        return "TOKEN EXPIRED"
+        
+    
     username : str = payload.get('sub')
     if username is None:
         raise credentials_exception
@@ -136,11 +142,24 @@ async def add_event(newEvent : schemas.EventCreate = Body(...) , db : Session = 
         return "event added successfully"
 
 @router.post('/delEvent')
-async def delete_event(* , event_id  : int = Body(...) , db: Session = Depends(get_db)):
-    event_to_delete = db.query(models.Event).get(event_id)
-    db.delete(event_to_delete)
-    db.commit()
-    return "event deleted"
+async def delete_event(* , event_id  : int = Body(...) , db: Session = Depends(get_db) , token : str = Body(...)):
+
+    try:
+        payload = jwt.decode(token , SECRET_KEY , algorithms=[ALGORITHM])
+    except:
+        return "TOKEN EXPIRED"
+    
+    username : str = payload.get('sub')
+    expiry_time = payload.get('exp')
+    expiry_time = datetime.fromtimestamp(expiry_time)
+    print((expiry_time - datetime.now()).seconds)
+    if username is None:
+        raise credentials_exception
+    else: 
+        event_to_delete = db.query(models.Event).get(event_id)
+        db.delete(event_to_delete)
+        db.commit()
+        return "event deleted"
 
 
 @router.post('/addLineup')
@@ -156,23 +175,35 @@ async def add_lineup(db : Session = Depends(get_db) , new_lineup : List[schemas.
     return "lineup added"
 
 @router.put('/updateState')
-async def set_new_state(new_event_state = Body(...) , db: Session = Depends(get_db)):
-    ev_id = new_event_state['event_id']
-    newState = new_event_state['state']
-    event_to_change = db.query(models.Event).filter(models.Event.id == ev_id )
-    print(event_to_change)
-    event_to_change.update({models.Event.state : newState} , synchronize_session = False)
-    db.commit()
-    return "state updated!"
+async def set_new_state(* , updated_event : schemas.UpdatedEvent = Body(...) , db: Session = Depends(get_db) , token : str = Body(...)):
+
+    try:
+        payload = jwt.decode(token , SECRET_KEY , algorithms=[ALGORITHM])
+    except:
+        return "TOKEN EXPIRED"
+        
+    username : str = payload.get('sub')
+    if username is None:
+        raise credentials_exception
+    else:   
+        ev_id = updated_event.id
+        newState = updated_event.state
+        newGalLink = updated_event.gallery_link
+        newPingLink = updated_event.ping_link
+        event_to_change = db.query(models.Event).filter(models.Event.id == ev_id )
+        print(event_to_change)
+        event_to_change.update({models.Event.state : newState , models.Event.gallery_link : newGalLink , models.Event.ping_link : newPingLink} , synchronize_session = False)
+        db.commit()
+        return "state updated!"
 
 @router.post('/addPhoto')
-async def add_photo(*, img_files : List[UploadFile] = File(...) , db : Session = Depends(get_db) , eventId : int = Body(...)):
+async def add_photo(*, img_files : List[UploadFile] = File(...) , db : Session = Depends(get_db) , eventId : int = Body(...) , photoLabel : str = Body(...)):
     for f in img_files:
-        data = await f.read(10)
+        data = await f.read()
         b64data = base64.b64encode(data)
         newImage = {
             'event_id' : eventId,
-            'label' : "test",
+            'label' : photoLabel,
             'image' : b64data
         }
         
@@ -185,9 +216,33 @@ async def add_photo(*, img_files : List[UploadFile] = File(...) , db : Session =
     return "photo saved in db"
 
 @router.post('/delPhoto')
-async def add_photo(photo_id  : int = Body(...) , db: Session = Depends(get_db)):
+async def delete_photo(*,photo_id  : int = Body(...) , db: Session = Depends(get_db) , token : str = Body(...)):
+
+    try:
+        payload = jwt.decode(token , SECRET_KEY , algorithms=[ALGORITHM])
+    except:
+        return "TOKEN EXPIRED"
+        
+    username : str = payload.get('sub')
+    expiry_time = payload.get('exp')
+    expiry_time = datetime.fromtimestamp(expiry_time)
+    print((expiry_time - datetime.now()).seconds)
+    if username is None:
+        raise credentials_exception
+    else:
+        image_to_delete = db.query(models.Photos).get(photo_id)
+        db.delete(image_to_delete)
+        db.commit()
+        return "image deleted"
+   
+
+@router.post('/confirm')
+async def confirm_delete(photo_id  : int = Body(...) , db: Session = Depends(get_db)):
     image_to_delete = db.query(models.Photos).get(photo_id)
     db.delete(image_to_delete)
     db.commit()
     return "photo has been deleted to db!"
+
+
+
